// mcp-servers/gatling/src/apiClientGenerated/gatlingEnterpriseSchemas.ts
export type SimpleDictionary = {
  [key: string]: string;
};
// ...
export type DistributionDto = {
  loadGenerators: LoadGeneratorDto[];
};

export type GlobalErrorRatioCriterionDto = {
  /**
   * @format int32
   */
  timeframeInSeconds: number;
  threshold: MaxPercentageThresholdDto;
};

export type GlobalResponseTimeCriterionDto = {
  /**
   * @format int32
   */
  timeframeInSeconds: number;
  threshold: MaxMillisecondsThresholdDto;
};

export type InstanceDto = {
  /**
   * @minimum 1
   * @format int32
   */
  count: number;
};

export type LoadGeneratorDto = {
  /**
   * Location ID
   */
  locationId: string;
  instance: InstanceDto;
  /**
   * @format int32
   */
  loadPercentage?: number;
};
// ...
export type MeaningfulTimeWindowDto = {
  /**
   * @format int32
   */
  rampUpSeconds: number;
  /**
   * @format int32
   */
  rampDownSeconds: number;
};
// ...
export type StopCriterionDto =
  | (Omit<StopCriterionDtoMeanCpu, "type"> & {
      type: "meanCpu";
    })
  | (Omit<StopCriterionDtoGlobalErrorRatio, "type"> & {
      type: "globalErrorRatio";
    })
  | (Omit<StopCriterionDtoGlobalResponseTime, "type"> & {
      type: "globalResponseTime";
    });

export type StopCriterionDtoGlobalErrorRatio = GlobalErrorRatioCriterionDto & StopCriterionDtoMixin;

export type StopCriterionDtoGlobalResponseTime = GlobalResponseTimeCriterionDto &
  StopCriterionDtoMixin;

export type StopCriterionDtoMeanCpu = MeanCpuCriterionDto & StopCriterionDtoMixin;
// ...
export type BuildFromSource = {
  /**
   * Source Repository ID
   */
  sourceRepositoryId: string;
  buildTool: BuildCommandDto;
  /**
   * Simulation class name
   */
  simulation: string;
  workingDirectory?: string;
  branch?: string;
};

export type CustomCommand = {
  command: string;
  /**
   * Package format
   */
  packageFormat: string;
};

export type DefaultCommand = Record<string, any>;

export type ExecutionDto = {
  meaningfulTimeWindow: MeaningfulTimeWindowDto;
  systemProperties: SimpleDictionary;
  environmentVariables: SimpleDictionary;
  /**
   * @default false
   */
  ignoreGlobalProperties: boolean;
  stopCriteria: StopCriterionDto[];
};

export type NoCode = {
  /**
   * Team ID
   */
  teamId: string;
};

export type Packaged = {
  /**
   * Package ID
   */
  packageId: string;
  /**
   * Simulation class name
   */
  simulation: string;
};

export type SourceDetails =
  | (Omit<SourceDetailsBuildFromSources, "type"> & {
      type: "build_from_sources";
    })
  | (Omit<SourceDetailsPackaged, "type"> & {
      type: "packaged";
    })
  | (Omit<SourceDetailsNoCode, "type"> & {
      type: "no_code";
    });

export type SourceDetailsBuildFromSources = BuildFromSource & SourceDetailsMixin;

export type SourceDetailsMixin = {
  type: string;
};

export type SourceDetailsNoCode = NoCode & SourceDetailsMixin;

export type SourceDetailsPackaged = Packaged & SourceDetailsMixin;
// ...
export type TestRequest = {
  /**
   * Non-empty trimmed string
   */
  name: string;
  distribution: DistributionDto;
  execution: ExecutionDto;
  source: SourceDetails;
};